status: Move status modules to separate names without depr. warnings
