+++ master/buildbot/status/__init__.py
@@ -1,16 +1,16 @@
-from buildbot.status import build
-from buildbot.status import builder
-from buildbot.status import buildrequest
-from buildbot.status import buildset
-from buildbot.status import master
+from buildbot.status import build_compat
+from buildbot.status import builder_compat
+from buildbot.status import buildrequest_compat
+from buildbot.status import buildset_compat
+from buildbot.status import master_compat
 
 # styles.Versioned requires this, as it keys the version numbers on the fully
 # qualified class name; see master/buildbot/test/regressions/test_unpickling.py
-build.BuildStatus.__module__ = 'buildbot.status.builder'
+build_compat.BuildStatus.__module__ = 'buildbot.status.builder'
 
 # add all of these classes to builder; this is a form of late binding to allow
 # circular module references among the status modules
-builder.BuildSetStatus = buildset.BuildSetStatus
-builder.Status = master.Status
-builder.BuildStatus = build.BuildStatus
-builder.BuildRequestStatus = buildrequest.BuildRequestStatus
+builder_compat.BuildSetStatus = buildset_compat.BuildSetStatus
+builder_compat.Status = master_compat.Status
+builder_compat.BuildStatus = build_compat.BuildStatus
+builder_compat.BuildRequestStatus = buildrequest_compat.BuildRequestStatus
