+++ master/buildbot/status/build_compat.py
@@ -20,12 +20,6 @@ from zope.interface import implementer
 
 from buildbot import interfaces
 from buildbot import util
-from buildbot.warnings import warn_deprecated
-
-warn_deprecated(
-    '0.9.0',
-    'buildbot.status.build has been deprecated, consume the buildbot.data APIs'
-)
 
 
 @implementer(interfaces.IBuildStatus, interfaces.IStatusEvent)
