+++ master/buildbot/status/buildrequest_compat.py
@@ -20,12 +20,6 @@ from zope.interface import implementer
 
 from buildbot import interfaces
 from buildbot.util.eventual import eventually
-from buildbot.warnings import warn_deprecated
-
-warn_deprecated(
-    '0.9.0',
-    'buildbot.status.buildrequest has been deprecated, consume the buildbot.data APIs'
-)
 
 
 @implementer(interfaces.IBuildRequestStatus)
