+++ master/buildbot/status/worker_compat.py
@@ -13,6 +13,7 @@
 #
 # Copyright Buildbot Team Members
 
+# this file
 
 import time
 
@@ -22,12 +23,6 @@ from buildbot import interfaces
 from buildbot.process.properties import Properties
 from buildbot.util import bytes2unicode
 from buildbot.util.eventual import eventually
-from buildbot.warnings import warn_deprecated
-
-warn_deprecated(
-    '0.9.0',
-    'buildbot.status.worker has been deprecated, consume the buildbot.data APIs'
-)
 
 
 @implementer(interfaces.IWorkerStatus)
