+++ master/buildbot/test/__init__.py
@@ -52,18 +52,20 @@ if parse_version(mock.__version__) < parse_version("0.8"):
 
 with assertProducesWarnings(DeprecatedApiWarning,
                             messages_patterns=[
+                                r" buildbot\.status\.base has been deprecated",
                                 r" buildbot\.status\.build has been deprecated",
                                 r" buildbot\.status\.buildrequest has been deprecated",
                                 r" buildbot\.status\.event has been deprecated",
-                                r" buildbot\.status\.worker has been deprecated",
                                 r" buildbot\.status\.buildset has been deprecated",
                                 r" buildbot\.status\.master has been deprecated",
-                                r" buildbot\.status\.base has been deprecated",
+                                r" buildbot\.status\.worker has been deprecated",
                             ]):
     import buildbot.status.base as _  # noqa
-
-with assertProducesWarning(DeprecatedApiWarning,
-                           message_pattern=r" buildbot\.status\.worker has been deprecated"):
+    import buildbot.status.build as _  # noqa
+    import buildbot.status.buildrequest as _  # noqa
+    import buildbot.status.event as _  # noqa
+    import buildbot.status.buildset as _  # noqa
+    import buildbot.status.master as _  # noqa
     import buildbot.status.worker as _  # noqa
 
 # All deprecated modules should be loaded, consider future warnings in tests as errors.
