+++ master/buildbot/test/regressions/test_oldpaths.py
@@ -14,6 +14,8 @@
 # Copyright Buildbot Team Members
 
 
+import warnings
+
 from twisted.trial import unittest
 
 from buildbot.warnings import DeprecatedApiWarning
@@ -96,30 +98,55 @@ class OldImportPaths(unittest.TestCase):
         assert SubunitShellCommand
 
     def test_status_builder_results(self):
-        # these symbols are now in buildbot.process.results, but lots of user
-        # code references them here:
-        from buildbot.status.builder import SUCCESS, WARNINGS, FAILURE, SKIPPED
-        from buildbot.status.builder import EXCEPTION, RETRY, Results
-        from buildbot.status.builder import worst_status
-        # reference the symbols to avoid failure from pyflakes
-        (SUCCESS, WARNINGS, FAILURE, SKIPPED, EXCEPTION, RETRY, Results,
-         worst_status)
+        # We can't reliable check for these warnings as they depend on whether the tests are
+        # run in parallel mode
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore", DeprecatedApiWarning)
+
+            # these symbols are now in buildbot.process.results, but lots of user
+            # code references them here:
+            from buildbot.status.builder import SUCCESS, WARNINGS, FAILURE, SKIPPED
+            from buildbot.status.builder import EXCEPTION, RETRY, Results
+            from buildbot.status.builder import worst_status
+            # reference the symbols to avoid failure from pyflakes
+            (SUCCESS, WARNINGS, FAILURE, SKIPPED, EXCEPTION, RETRY, Results,
+             worst_status)
 
     def test_status_builder_BuildSetStatus(self):
-        from buildbot.status.builder import BuildSetStatus
-        assert BuildSetStatus
+        # We can't reliable check for these warnings as they depend on whether the tests are
+        # run in parallel mode
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore", DeprecatedApiWarning)
+
+            from buildbot.status.builder import BuildSetStatus
+            assert BuildSetStatus
 
     def test_status_builder_Status(self):
-        from buildbot.status.builder import Status
-        assert Status
+        # We can't reliable check for these warnings as they depend on whether the tests are
+        # run in parallel mode
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore", DeprecatedApiWarning)
+
+            from buildbot.status.builder import Status
+            assert Status
 
     def test_status_builder_Event(self):
-        from buildbot.status.builder import Event
-        assert Event
+        # We can't reliable check for these warnings as they depend on whether the tests are
+        # run in parallel mode
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore", DeprecatedApiWarning)
+
+            from buildbot.status.builder import Event
+            assert Event
 
     def test_status_builder_BuildStatus(self):
-        from buildbot.status.builder import BuildStatus
-        assert BuildStatus
+        # We can't reliable check for these warnings as they depend on whether the tests are
+        # run in parallel mode
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore", DeprecatedApiWarning)
+
+            from buildbot.status.builder import BuildStatus
+            assert BuildStatus
 
     def test_steps_source_Source(self):
         from buildbot.steps.source import Source
